Fix permission error surrounding kibana.pid
Constant deletion/creation of this file was causing kibana to
fail in certain scenarios. This change alters the kibana service
to always ensure the pid file exists and is owned by kibana
every time the service starts (or is restarted).

Previously we attempted to address this during ansible provisioning.
This solution did not protected against all scenarios, however.